Try failing pytest

+++ tests/test_lexer.py
@@ -4,7 +4,8 @@ from pyxpp.lexer import tokenize
 
 def test_literals():
     literals = '1.0 4 0 \n'
-    expected = ['FLOAT', 'INTEGER', 'INTEGER', 'NEWLINE']
+    # expected = ['FLOAT', 'INTEGER', 'INTEGER', 'NEWLINE']
+    expected = ['INTEGER', 'INTEGER', 'INTEGER', 'NEWLINE']
     tokens = tokenize(literals)
     assert [token.type for token in tokens] == expected
